create

+++ src/events/eventsController.ts
@@ -14,7 +14,8 @@ import {
   ValidationPipe,
   NotFoundException,
   Query,
-  UsePipes
+  UsePipes,
+  ForbiddenException
 } from '@nestjs/common';
 import { CreateEventDto } from './input/create-event.dto';
 import { Event } from './event.entity';
@@ -24,14 +25,15 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Attendee } from './attendee.entity';
 import { EventsService } from './events.service';
 import { ListEvents } from './input/list.events';
+import { UseGuards } from '@nestjs/common';
+import { AuthGuardJwt } from "src/auth/auth-guard.jwt";
+import { CurrentUser } from 'src/auth/current-user.decorator';
+import { User } from 'src/auth/user.entity';
+
 @Controller('/events')
 export class EventsController {
   private readonly logger = new Logger(EventsController.name);
   constructor(
-    @InjectRepository(Event)
-    private repository: Repository<Event>,
-    @InjectRepository(Attendee)
-    private attendeeRepository: Repository<Attendee>,
     private readonly eventsService:EventsService,
   ) {}
 
@@ -54,7 +56,7 @@ export class EventsController {
   }
 
   
-  @Get("/practicee")
+ /* @Get("/practicee")
   async practice2() {
   /*   const event= new Event();
     event.id=1
@@ -72,14 +74,14 @@ export class EventsController {
 
     if (!event) {
       throw new NotFoundException();
-    } */ 
+    }  
 
     return await this.repository.createQueryBuilder('e')
     .select(['e.id', 'e.name'])
     .orderBy('e.id', 'DESC')
     .take(3) 
     .getMany();
-  }
+  }*/
 
 
 
@@ -94,7 +96,7 @@ export class EventsController {
     return event
   }
 
-  @Get('/practice')
+  /*@Get('/practice')
   async practice() {
     return await this.repository.find({
       select: ['id', 'when'],
@@ -113,40 +115,55 @@ export class EventsController {
       },
     });
   }
-
+*/
 
 
   @Patch(':id')
-  async update(@Param('id') id, @Body() input: UpdateEventDto) {
+  @UseGuards(AuthGuardJwt)
+    async update(
+    @Param('id') id, 
+    @Body() input: UpdateEventDto,
+    @CurrentUser() user: User) {
     
-    const event = await this.repository.findOneBy({id:id});
+      const event = await this.eventsService.getEvent(id);
+      if (!event) {
+      throw new NotFoundException()
+      
+    }
+    if (event.organizerId!==user.id) {
+      throw new ForbiddenException(null,'you are not authorized to change this event')
+      
+    }
+
+    return await this.eventsService.updateEvent(event,input)
+    
+  }
+
+  @Post()
+  @UseGuards(AuthGuardJwt)
+  async create(
+    @Body() input: CreateEventDto,
+    @CurrentUser() user: User
+  ) {
+    return await this.eventsService.createEvent(input, user);
+  }
+
+
+  @Delete(':id')
+  @UseGuards(AuthGuardJwt)
+  @HttpCode(204)
+  async remove(@Param('id') id, @CurrentUser() user: User  ) {
+    const event = await this.eventsService.getEvent(id);
     if (!event) {
       throw new NotFoundException()
       
     }
-    return await this.repository.save({
-      ...event,
-      ...input,
-      when: input.when ? new Date(input.when) : event.when
-    });
-
-  }
-
-  @Post()
-  async create(@Body() input: CreateEventDto) {
-    return await this.repository.save({
-      ...input,
-      when: new Date(input.when),
-    });
-  }
-
-  @Delete(':id')
-  @HttpCode(204)
-  async remove(@Param('id') id) {
-    const result = await this.eventsService.deleteEvent(id)
-    if (result.affected !== 1) {
-      throw new NotFoundException()
+    if (event.organizerId!==user.id) {
+      throw new ForbiddenException(null,'you are not authorized to delet this event')
+      
     }
+     await this.eventsService.deleteEvent(id)
+ 
   //  return result
